+++ Nickel/Framework/Implementations/Content/ModArtifacts.cs
@@ -1,5 +1,6 @@
 using System;
 using System.Collections.Generic;
+using System.Linq;
 
 namespace Nickel;
 
@@ -32,4 +33,26 @@ internal sealed class ModArtifacts(
 		this.RegisteredArtifactStorage[name] = entry;
 		return entry;
 	}
+}
+
+internal sealed class VanillaModArtifacts(
+	Func<ArtifactManager> artifactManagerProvider
+) : IModArtifacts
+{
+	private readonly Lazy<Dictionary<string, IArtifactEntry>> LazyRegisteredArtifacts = new(() => DB.artifacts.Where(kvp => kvp.Value.Assembly == typeof(Card).Assembly).ToDictionary(kvp => kvp.Key, kvp => artifactManagerProvider().LookupByArtifactType(kvp.Value)!));
+	
+	public IReadOnlyDictionary<string, IArtifactEntry> RegisteredArtifacts
+		=>  this.LazyRegisteredArtifacts.Value;
+	
+	public IArtifactEntry? LookupByArtifactType(Type artifactType)
+		=> artifactManagerProvider().LookupByArtifactType(artifactType);
+
+	public IArtifactEntry? LookupByUniqueName(string uniqueName)
+		=> artifactManagerProvider().LookupByUniqueName(uniqueName);
+
+	public IArtifactEntry RegisterArtifact(ArtifactConfiguration configuration)
+		=> throw new NotSupportedException();
+
+	public IArtifactEntry RegisterArtifact(string name, ArtifactConfiguration configuration)
+		=> throw new NotSupportedException();
 }