+++ Nickel/Framework/Implementations/Content/ModCharactersV2.cs
@@ -1,5 +1,6 @@
 using System;
 using System.Collections.Generic;
+using System.Linq;
 
 namespace Nickel;
 
@@ -58,4 +59,49 @@ internal sealed class ModCharactersV2(
 		this.RegisteredNonPlayableCharacterStorage[name] = entry;
 		return entry;
 	}
+}
+
+internal sealed class VanillaModCharactersV2(
+	IModManifest modManifest,
+	Func<CharacterManager> characterManagerProvider
+) : IModCharactersV2
+{
+	private readonly Lazy<Dictionary<string, IPlayableCharacterEntryV2>> LazyRegisteredPlayableCharacters = new(
+		() => NewRunOptions.allChars
+			.Select(d => characterManagerProvider().LookupByDeckV2(d))
+			.Where(e => e?.ModOwner == modManifest)
+			.ToDictionary(e => e!.UniqueName, e => e!)
+	);
+	
+	public IReadOnlyDictionary<string, IPlayableCharacterEntryV2> RegisteredPlayableCharacters
+		=> this.LazyRegisteredPlayableCharacters.Value;
+	
+	// TODO: maybe implement one day
+	public IReadOnlyDictionary<string, INonPlayableCharacterEntryV2> RegisteredNonPlayableCharacters
+		=> throw new NotImplementedException();
+	
+	// TODO: maybe implement one day
+	public IReadOnlyDictionary<string, ICharacterAnimationEntryV2> RegisteredCharacterAnimations
+		=> throw new NotImplementedException();
+	
+	public ICharacterAnimationEntryV2 RegisterCharacterAnimation(CharacterAnimationConfigurationV2 configuration)
+		=> throw new NotSupportedException();
+
+	public ICharacterAnimationEntryV2 RegisterCharacterAnimation(string name, CharacterAnimationConfigurationV2 configuration)
+		=> throw new NotSupportedException();
+
+	public IPlayableCharacterEntryV2? LookupByDeck(Deck deck)
+		=> characterManagerProvider().LookupByDeckV2(deck);
+
+	public ICharacterEntryV2? LookupByCharacterType(string characterType)
+		=> characterManagerProvider().LookupByCharacterTypeV2(characterType);
+
+	public ICharacterEntryV2? LookupByUniqueName(string uniqueName)
+		=> characterManagerProvider().LookupByUniqueNameV2(uniqueName);
+
+	public IPlayableCharacterEntryV2 RegisterPlayableCharacter(string name, PlayableCharacterConfigurationV2 configuration)
+		=> throw new NotSupportedException();
+
+	public INonPlayableCharacterEntryV2 RegisterNonPlayableCharacter(string name, NonPlayableCharacterConfigurationV2 configuration)
+		=> throw new NotSupportedException();
 }