+++ Nickel/Framework/Implementations/Content/ModDecks.cs
@@ -1,5 +1,6 @@
 using System;
 using System.Collections.Generic;
+using System.Linq;
 
 namespace Nickel;
 
@@ -25,4 +26,23 @@ internal sealed class ModDecks(
 		this.RegisteredDeckStorage[name] = entry;
 		return entry;
 	}
+}
+
+internal sealed class VanillaModDecks(
+	Func<DeckManager> deckManagerProvider
+) : IModDecks
+{
+	private readonly Lazy<Dictionary<string, IDeckEntry>> LazyRegisteredDecks = new(() => Enum.GetValues<Deck>().Select(d => deckManagerProvider().LookupByDeck(d)!).ToDictionary(e => e.Deck.Key()));
+	
+	public IReadOnlyDictionary<string, IDeckEntry> RegisteredDecks
+		=> this.LazyRegisteredDecks.Value;
+	
+	public IDeckEntry? LookupByDeck(Deck deck)
+		=> deckManagerProvider().LookupByDeck(deck);
+
+	public IDeckEntry? LookupByUniqueName(string uniqueName)
+		=> deckManagerProvider().LookupByUniqueName(uniqueName);
+
+	public IDeckEntry RegisterDeck(string name, DeckConfiguration configuration)
+		=> throw new NotSupportedException();
 }